refactor: reorganize plugins acording to new standard
